I decided to temporarily remove mixins.py as unnecessary.
I have finalized the logic of AddNote, all that remains is to set up a beautiful display of all records.
Temporarily abandoned the idea of storing the logic for creating query queries in a separate mixin. It would be more practical to inherit from the abstract handle and then re-inherit from the class with the query logic/